+++ server/src/bruFiles/diagnostics/requestFiles/getMetaBlockSpecificDiagnostics.ts
@@ -9,6 +9,9 @@ import {
     DictionaryBlock,
     isDictionaryBlockArrayField,
     DictionaryBlockArrayField,
+    getMetaBlockMandatoryKeys,
+    getMetaBlockOptionalKeys,
+    BrunoFileType,
 } from "@global_shared";
 import { checkNoDuplicateKeysAreDefinedForDictionaryBlock } from "../shared/checks/singleBlocks/checkNoDuplicateKeysAreDefinedForDictionaryBlock";
 import { checkNoKeysAreMissingForDictionaryBlock } from "../shared/checks/singleBlocks/checkNoKeysAreMissingForDictionaryBlock";
@@ -33,12 +36,17 @@ export async function getMetaBlockSpecificDiagnostics(
     documentHelper: TextDocumentHelper,
     metaBlock: DictionaryBlock,
 ): Promise<(DiagnosticWithCode | undefined)[]> {
-    const mandatoryBlockKeys = [
-        MetaBlockKey.Name,
-        MetaBlockKey.Sequence,
-        MetaBlockKey.Type,
-    ];
-    const optionalBlockKeys = [MetaBlockKey.Tags];
+    const mandatoryBlockKeys = getMetaBlockMandatoryKeys(
+        BrunoFileType.RequestFile,
+    );
+    const optionalBlockKeys = getMetaBlockOptionalKeys(
+        BrunoFileType.RequestFile,
+    );
+
+    if (!mandatoryBlockKeys || !optionalBlockKeys) {
+        return [];
+    }
+
     const typeFields = metaBlock.content.filter(
         ({ key, disabled }) => key == MetaBlockKey.Type && !disabled,
     );