+++ shared/languageUtils/commonBlocks/metaBlock/getMandatoryAndOptionalKeys.ts
@@ -1,0 +1,15 @@
+import { BrunoFileType, MetaBlockKey } from "../../..";
+
+export function getMetaBlockMandatoryKeys(fileType: BrunoFileType) {
+    return fileType == BrunoFileType.RequestFile
+        ? [MetaBlockKey.Name, MetaBlockKey.Sequence, MetaBlockKey.Type]
+        : fileType == BrunoFileType.FolderSettingsFile
+          ? [MetaBlockKey.Name, MetaBlockKey.Sequence]
+          : undefined;
+}
+
+export function getMetaBlockOptionalKeys(fileType: BrunoFileType) {
+    return fileType == BrunoFileType.RequestFile
+        ? [MetaBlockKey.Tags]
+        : undefined;
+}