+++ shared/languageUtils/commonBlocks/settingsBlock/getMandatoryKeys.ts
@@ -1,0 +1,5 @@
+import { SettingsBlockKey } from "../../..";
+
+export function getMandatoryKeys() {
+    return Object.values(SettingsBlockKey);
+}