+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Fixes / Minor Improvements
 
 - Improve grouping and sorting of context menu items in collection explorer.
+- Provide completions for keys in all dictionary blocks, except for auth blocks of type OAuth2.
 
 ## [0.22.1] (2026-04-02)